# Optimize load csv