--- a/src/mlpack/core/data/load_csv.hpp
+++ b/src/mlpack/core/data/load_csv.hpp
@@ -33,7 +33,7 @@ public:
     extension(Extension(file)),
     fatalIfOpenFail(fatal),
     fileName(std::move(file)),
-    inFile(file)
+    inFile(fileName)
   {
     CanOpen();
   }