… HISTORY.md
… doc/tutorials/range_search/range_search.txt
… src/mlpack/core/data/CMakeLists.txt
… src/mlpack/core/dists/gaussian_distribution.cpp
… src/mlpack/core/tree/ballbound.hpp
… src/mlpack/core/tree/ballbound_impl.hpp
… src/mlpack/core/tree/binary_space_tree/typedef.hpp
… src/mlpack/core/tree/rectangle_tree/r_star_tree_split_impl.hpp
… src/mlpack/core/tree/rectangle_tree/r_tree_split_impl.hpp
… src/mlpack/core/tree/rectangle_tree/rectangle_tree_impl.hpp
… src/mlpack/core/tree/rectangle_tree/x_tree_split_impl.hpp
… src/mlpack/core/tree/traversal_info.hpp
… src/mlpack/methods/CMakeLists.txt
… src/mlpack/methods/ann/activation_functions/logistic_function.hpp
… src/mlpack/methods/ann/layer/constant_layer.hpp
… src/mlpack/methods/ann/layer/dropconnect_layer.hpp
… src/mlpack/methods/ann/layer/empty_layer.hpp
… src/mlpack/methods/ann/layer/glimpse_layer.hpp
… src/mlpack/methods/ann/layer/hard_tanh_layer.hpp
… src/mlpack/methods/ann/layer/leaky_relu_layer.hpp
… src/mlpack/methods/ann/layer/log_softmax_layer.hpp
… src/mlpack/methods/ann/layer/multiply_constant_layer.hpp
… src/mlpack/methods/ann/layer/negative_log_likelihood_layer.hpp
… src/mlpack/methods/ann/layer/reinforce_normal_layer.hpp
… src/mlpack/methods/ann/layer/vr_class_reward_layer.hpp
… src/mlpack/methods/emst/dtb_rules.hpp
… src/mlpack/methods/fastmks/fastmks_rules.hpp
… src/mlpack/methods/hmm/hmm_impl.hpp
… src/mlpack/methods/hmm/hmm_train_main.cpp
… src/mlpack/methods/kmeans/dual_tree_kmeans_rules.hpp