Hi, I use boost::spirit to implement the csv parser, it is more memory efficient and faster.

parse file with 1 million lines, 39796KByte

spirit version :

transpose : 2151 msec
non transpose : 4073 msec

old version :

transpose : 9616 msec
non transpose : 10131 msec

non transpose version is slower, I guess it is because arma::Mat is column wise.

Upload for code reviews, haven't integrate it into the load function and run the test cases yet.

ps : Single thread only, do not know multi-thread can make performance become better or worse, DataSetInfo is not a lock free data structure. If we want to utilize the power of multi-thread, I think we could read a bunch of string into the vector, create thread pool and DataSetInfo vectors, merge the DataSetInfo at last.