@keonkim 2016-06-05
@stereomatchingkiss hmm... did you amend it with the merge commit?