@stereomatchingkiss 2016-06-05
>@stereomatchingkiss hmm... did you amend it with the merge commit?

No, I only merge pull request(#650) one time.

>Someone force pushed to the repo and I am in the process of fixing it. That will fix this problem.

Thanks for the fix.