@stereomatchingkiss 2016-06-06
@rcurtin Would it make things easier to deal with if I delete this branch and open a new one?